requirements.txt
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
flask_script
PyJWT==1.7.1
psycopg2==2.8.4
